Util: Add the JSONReader helper
This patch adds a `JSONReader()` helper method to facilitate passing eg. a `map[string]string` as the body,
without requiring the user to manually encode it into JSON and create eg. a `strings.Reader`:

  res, err = es.Search(
		es.Search.WithIndex("test"),
		es.Search.WithBody(esutil.JSONReader(map[string]string{"query": {"match": {"title": "test"}}})),
	)

The package defines the `JSONEncoder` interface, so it's possible for the outside code to provide
an object implementing it, in order to use eg. a third-party JSON package. The default encoder
is "encoding/json".